+++ src/angular/ast-utils.ts
@@ -1,12 +1,27 @@
-/**
- * @license
- * Copyright Google Inc. All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-import * as ts from 'typescript';
-import { Change, InsertChange, NoopChange } from './change';
+import {
+  insertImport as originalInsertImport,
+  findNodes as originalFindNodes,
+  findNode as originalFindNode,
+  getSourceNodes as originalGetSourceNodes,
+  insertAfterLastOccurrence as originalInsertAfterLastOccurrence,
+  getContentOfKeyLiteral as originalGetContentOfKeyLiteral,
+  getFirstNgModuleName as originalGetFirstNgModuleName,
+  getDecoratorMetadata as originalGetDecoratorMetadata,
+  getMetadataField as originalGetMetadataField,
+  getRouterModuleDeclaration as originalGetRouterModuleDeclaration,
+  addSymbolToNgModuleMetadata as originalAddSymbolToNgModuleMetadata,
+  addDeclarationToModule as originalAddDeclarationToModule,
+  addImportToModule as originalAddImportToModule,
+  addProviderToModule as originalAddProviderToModule,
+  addExportToModule as originalAddExportToModule,
+  addBootstrapToModule as originalAddBootstrapToModule,
+  addEntryComponentToModule as originalAddEntryComponentToModule,
+  addRouteDeclarationToModule as originalAddRouteDeclarationToModule,
+  isImported as originalIsImported
+} from '@schematics/angular/utility/ast-utils';
+import * as ts from '@schematics/angular/third_party/github.com/Microsoft/TypeScript/lib/typescript';
+
+import { Change } from './change';
 
 /**
  * Add Import `import { symbolName } from fileName` if the import doesn't exit
@@ -24,67 +39,7 @@ export function insertImport(
   fileName: string,
   isDefault = false
 ): Change {
-  const rootNode = source;
-  const allImports = findNodes(rootNode, ts.SyntaxKind.ImportDeclaration);
-
-  // get nodes that map to import statements from the file fileName
-  const relevantImports = allImports.filter(node => {
-    // StringLiteral of the ImportDeclaration is the import file (fileName in this case).
-    const importFiles = node
-      .getChildren()
-      .filter(child => child.kind === ts.SyntaxKind.StringLiteral)
-      .map(n => (n as ts.StringLiteral).text);
-
-    return importFiles.filter(file => file === fileName).length === 1;
-  });
-
-  if (relevantImports.length > 0) {
-    let importsAsterisk = false;
-    // imports from import file
-    const imports: ts.Node[] = [];
-    relevantImports.forEach(n => {
-      Array.prototype.push.apply(imports, findNodes(n, ts.SyntaxKind.Identifier));
-      if (findNodes(n, ts.SyntaxKind.AsteriskToken).length > 0) {
-        importsAsterisk = true;
-      }
-    });
-
-    // if imports * from fileName, don't add symbolName
-    if (importsAsterisk) {
-      return new NoopChange();
-    }
-
-    const importTextNodes = imports.filter(n => (n as ts.Identifier).text === symbolName);
-
-    // insert import if it's not there
-    if (importTextNodes.length === 0) {
-      const fallbackPos =
-        findNodes(relevantImports[0], ts.SyntaxKind.CloseBraceToken)[0].getStart() ||
-        findNodes(relevantImports[0], ts.SyntaxKind.FromKeyword)[0].getStart();
-
-      return insertAfterLastOccurrence(imports, `, ${symbolName}`, fileToEdit, fallbackPos);
-    }
-
-    return new NoopChange();
-  }
-
-  // no such import declaration exists
-  const useStrict = findNodes(rootNode, ts.SyntaxKind.StringLiteral).filter(
-    (n: ts.StringLiteral) => n.text === 'use strict'
-  );
-  let fallbackPos = 0;
-  if (useStrict.length > 0) {
-    fallbackPos = useStrict[0].end;
-  }
-  const open = isDefault ? '' : '{ ';
-  const close = isDefault ? '' : ' }';
-  // if there are no imports or 'use strict' statement, insert import at beginning of file
-  const insertAtBeginning = allImports.length === 0 && useStrict.length === 0;
-  const separator = insertAtBeginning ? '' : ';\n';
-  const toInsert =
-    `${separator}import ${open}${symbolName}${close}` + ` from '${fileName}'${insertAtBeginning ? ';\n' : ''}`;
-
-  return insertAfterLastOccurrence(allImports, toInsert, fileToEdit, fallbackPos, ts.SyntaxKind.StringLiteral);
+  return originalInsertImport(source, fileToEdit, symbolName, fileName, isDefault);
 }
 
 /**
@@ -92,34 +47,12 @@ export function insertImport(
  * @param node
  * @param kind
  * @param max The maximum number of items to return.
+ * @param recursive Continue looking for nodes of kind recursive until end
+ * the last child even when node of kind has been found.
  * @return all nodes of kind, or [] if none is found
  */
-export function findNodes(node: ts.Node, kind: ts.SyntaxKind, max = Infinity): ts.Node[] {
-  if (!node || max == 0) {
-    return [];
-  }
-
-  const arr: ts.Node[] = [];
-  if (node.kind === kind) {
-    arr.push(node);
-    max--;
-  }
-  if (max > 0) {
-    for (const child of node.getChildren()) {
-      findNodes(child, kind, max).forEach(node => {
-        if (max > 0) {
-          arr.push(node);
-        }
-        max--;
-      });
-
-      if (max <= 0) {
-        break;
-      }
-    }
-  }
-
-  return arr;
+export function findNodes(node: ts.Node, kind: ts.SyntaxKind, max = Infinity, recursive = false): ts.Node[] {
+  return originalFindNodes(node, kind, max, recursive);
 }
 
 /**
@@ -128,43 +61,11 @@ export function findNodes(node: ts.Node, kind: ts.SyntaxKind, max = Infinity): t
  * @returns {Observable<ts.Node>} An observable of all the nodes in the source.
  */
 export function getSourceNodes(sourceFile: ts.SourceFile): ts.Node[] {
-  const nodes: ts.Node[] = [sourceFile];
-  const result = [];
-
-  while (nodes.length > 0) {
-    const node = nodes.shift();
-
-    if (node) {
-      result.push(node);
-      if (node.getChildCount(sourceFile) >= 0) {
-        nodes.unshift(...node.getChildren());
-      }
-    }
-  }
-
-  return result;
+  return originalGetSourceNodes(sourceFile);
 }
 
 export function findNode(node: ts.Node, kind: ts.SyntaxKind, text: string): ts.Node | null {
-  if (node.kind === kind && node.getText() === text) {
-    // throw new Error(node.getText());
-    return node;
-  }
-
-  let foundNode: ts.Node | null = null;
-  ts.forEachChild(node, childNode => {
-    foundNode = foundNode || findNode(childNode, kind, text);
-  });
-
-  return foundNode;
-}
-
-/**
- * Helper for sorting nodes.
- * @return function to sort nodes in increasing order of position in sourceFile
- */
-function nodesByPosition(first: ts.Node, second: ts.Node): number {
-  return first.getStart() - second.getStart();
+  return originalFindNode(node, kind, text);
 }
 
 /**
@@ -187,133 +88,15 @@ export function insertAfterLastOccurrence(
   fallbackPos: number,
   syntaxKind?: ts.SyntaxKind
 ): Change {
-  // sort() has a side effect, so make a copy so that we won't overwrite the parent's object.
-  let lastItem = [...nodes].sort(nodesByPosition).pop();
-  if (!lastItem) {
-    throw new Error();
-  }
-  if (syntaxKind) {
-    lastItem = findNodes(lastItem, syntaxKind)
-      .sort(nodesByPosition)
-      .pop();
-  }
-  if (!lastItem && fallbackPos == undefined) {
-    throw new Error(`tried to insert ${toInsert} as first occurence with no fallback position`);
-  }
-  const lastItemPosition: number = lastItem ? lastItem.getEnd() : fallbackPos;
-
-  return new InsertChange(file, lastItemPosition, toInsert);
+  return originalInsertAfterLastOccurrence(nodes, toInsert, file, fallbackPos, syntaxKind);
 }
 
 export function getContentOfKeyLiteral(_source: ts.SourceFile, node: ts.Node): string | null {
-  if (node.kind == ts.SyntaxKind.Identifier) {
-    return (node as ts.Identifier).text;
-  } else if (node.kind == ts.SyntaxKind.StringLiteral) {
-    return (node as ts.StringLiteral).text;
-  } else {
-    return null;
-  }
-}
-
-/**
- * @private
- */
-function _angularImportsFromNode(node: ts.ImportDeclaration, _sourceFile: ts.SourceFile): { [name: string]: string } {
-  const ms = node.moduleSpecifier;
-  let modulePath: string;
-  switch (ms.kind) {
-    case ts.SyntaxKind.StringLiteral:
-      modulePath = (ms as ts.StringLiteral).text;
-      break;
-    default:
-      return {};
-  }
-
-  if (!modulePath.startsWith('@angular/')) {
-    return {};
-  }
-
-  if (node.importClause) {
-    if (node.importClause.name) {
-      // This is of the form `import Name from 'path'`. Ignore.
-      return {};
-    } else if (node.importClause.namedBindings) {
-      const nb = node.importClause.namedBindings;
-      if (nb.kind == ts.SyntaxKind.NamespaceImport) {
-        // This is of the form `import * as name from 'path'`. Return `name.`.
-        return {
-          [(nb as ts.NamespaceImport).name.text + '.']: modulePath
-        };
-      } else {
-        // This is of the form `import {a,b,c} from 'path'`
-        const namedImports = nb as ts.NamedImports;
-
-        return namedImports.elements
-          .map((is: ts.ImportSpecifier) => (is.propertyName ? is.propertyName.text : is.name.text))
-          .reduce((acc: { [name: string]: string }, curr: string) => {
-            acc[curr] = modulePath;
-
-            return acc;
-          }, {});
-      }
-    }
-
-    return {};
-  } else {
-    // This is of the form `import 'path';`. Nothing to do.
-    return {};
-  }
+  return originalGetContentOfKeyLiteral(_source, node);
 }
 
 export function getDecoratorMetadata(source: ts.SourceFile, identifier: string, module: string): ts.Node[] {
-  const angularImports: { [name: string]: string } = findNodes(source, ts.SyntaxKind.ImportDeclaration)
-    .map((node: ts.ImportDeclaration) => _angularImportsFromNode(node, source))
-    .reduce((acc: { [name: string]: string }, current: { [name: string]: string }) => {
-      for (const key of Object.keys(current)) {
-        acc[key] = current[key];
-      }
-
-      return acc;
-    }, {});
-
-  return getSourceNodes(source)
-    .filter(node => {
-      return (
-        node.kind == ts.SyntaxKind.Decorator && (node as ts.Decorator).expression.kind == ts.SyntaxKind.CallExpression
-      );
-    })
-    .map(node => (node as ts.Decorator).expression as ts.CallExpression)
-    .filter(expr => {
-      if (expr.expression.kind == ts.SyntaxKind.Identifier) {
-        const id = expr.expression as ts.Identifier;
-
-        return id.getFullText(source) == identifier && angularImports[id.getFullText(source)] === module;
-      } else if (expr.expression.kind == ts.SyntaxKind.PropertyAccessExpression) {
-        // This covers foo.NgModule when importing * as foo.
-        const paExpr = expr.expression as ts.PropertyAccessExpression;
-        // If the left expression is not an identifier, just give up at that point.
-        if (paExpr.expression.kind !== ts.SyntaxKind.Identifier) {
-          return false;
-        }
-
-        const id = paExpr.name.text;
-        const moduleId = (paExpr.expression as ts.Identifier).getText(source);
-
-        return id === identifier && angularImports[moduleId + '.'] === module;
-      }
-
-      return false;
-    })
-    .filter(expr => expr.arguments[0] && expr.arguments[0].kind == ts.SyntaxKind.ObjectLiteralExpression)
-    .map(expr => expr.arguments[0] as ts.ObjectLiteralExpression);
-}
-
-function findClassDeclarationParent(node: ts.Node): ts.ClassDeclaration | undefined {
-  if (ts.isClassDeclaration(node)) {
-    return node;
-  }
-
-  return node.parent && findClassDeclarationParent(node.parent);
+  return originalGetDecoratorMetadata(source, identifier, module);
 }
 
 /**
@@ -323,21 +106,11 @@ function findClassDeclarationParent(node: ts.Node): ts.ClassDeclaration | undefi
  * @returns the name of the first @NgModule, or `undefined` if none is found
  */
 export function getFirstNgModuleName(source: ts.SourceFile): string | undefined {
-  // First, find the @NgModule decorators.
-  const ngModulesMetadata = getDecoratorMetadata(source, 'NgModule', '@angular/core');
-  if (ngModulesMetadata.length === 0) {
-    return undefined;
-  }
+  return originalGetFirstNgModuleName(source);
+}
 
-  // Then walk parent pointers up the AST, looking for the ClassDeclaration parent of the NgModule
-  // metadata.
-  const moduleClass = findClassDeclarationParent(ngModulesMetadata[0]);
-  if (!moduleClass || !moduleClass.name) {
-    return undefined;
-  }
-
-  // Get the class name of the module ClassDeclaration.
-  return moduleClass.name.text;
+export function getMetadataField(node: ts.ObjectLiteralExpression, metadataField: string): ts.ObjectLiteralElement[] {
+  return originalGetMetadataField(node, metadataField);
 }
 
 export function addSymbolToNgModuleMetadata(
@@ -347,136 +120,7 @@ export function addSymbolToNgModuleMetadata(
   symbolName: string,
   importPath: string | null = null
 ): Change[] {
-  const nodes = getDecoratorMetadata(source, 'NgModule', '@angular/core');
-  let node: any = nodes[0]; // tslint:disable-line:no-any
-
-  // Find the decorator declaration.
-  if (!node) {
-    return [];
-  }
-
-  // Get all the children property assignment of object literals.
-  const matchingProperties: ts.ObjectLiteralElement[] = (node as ts.ObjectLiteralExpression).properties
-    .filter(prop => prop.kind == ts.SyntaxKind.PropertyAssignment)
-    // Filter out every fields that's not "metadataField". Also handles string literals
-    // (but not expressions).
-    .filter((prop: ts.PropertyAssignment) => {
-      const name = prop.name;
-      switch (name.kind) {
-        case ts.SyntaxKind.Identifier:
-          return (name as ts.Identifier).getText(source) == metadataField;
-        case ts.SyntaxKind.StringLiteral:
-          return (name as ts.StringLiteral).text == metadataField;
-      }
-
-      return false;
-    });
-
-  // Get the last node of the array literal.
-  if (!matchingProperties) {
-    return [];
-  }
-  if (matchingProperties.length == 0) {
-    // We haven't found the field in the metadata declaration. Insert a new field.
-    const expr = node as ts.ObjectLiteralExpression;
-    let position: number;
-    let toInsert: string;
-    if (expr.properties.length == 0) {
-      position = expr.getEnd() - 1;
-      toInsert = `  ${metadataField}: [${symbolName}]\n`;
-    } else {
-      node = expr.properties[expr.properties.length - 1];
-      position = node.getEnd();
-      // Get the indentation of the last element, if any.
-      const text = node.getFullText(source);
-      const matches = text.match(/^\r?\n\s*/);
-      if (matches.length > 0) {
-        toInsert = `,${matches[0]}${metadataField}: [${symbolName}]`;
-      } else {
-        toInsert = `, ${metadataField}: [${symbolName}]`;
-      }
-    }
-    if (importPath !== null) {
-      return [
-        new InsertChange(ngModulePath, position, toInsert),
-        insertImport(source, ngModulePath, symbolName.replace(/\..*$/, ''), importPath)
-      ];
-    } else {
-      return [new InsertChange(ngModulePath, position, toInsert)];
-    }
-  }
-  const assignment = matchingProperties[0] as ts.PropertyAssignment;
-
-  // If it's not an array, nothing we can do really.
-  if (assignment.initializer.kind !== ts.SyntaxKind.ArrayLiteralExpression) {
-    return [];
-  }
-
-  const arrLiteral = assignment.initializer as ts.ArrayLiteralExpression;
-  if (arrLiteral.elements.length == 0) {
-    // Forward the property.
-    node = arrLiteral;
-  } else {
-    node = arrLiteral.elements;
-  }
-
-  if (!node) {
-    console.log('No app module found. Please add your new class to your component.');
-
-    return [];
-  }
-
-  if (Array.isArray(node)) {
-    const nodeArray = (node as {}) as Array<ts.Node>;
-    const symbolsArray = nodeArray.map(node => node.getText());
-    if (symbolsArray.includes(symbolName)) {
-      return [];
-    }
-
-    node = node[node.length - 1];
-  }
-
-  let toInsert: string;
-  let position = node.getEnd();
-  if (node.kind == ts.SyntaxKind.ObjectLiteralExpression) {
-    // We haven't found the field in the metadata declaration. Insert a new
-    // field.
-    const expr = node as ts.ObjectLiteralExpression;
-    if (expr.properties.length == 0) {
-      position = expr.getEnd() - 1;
-      toInsert = `  ${metadataField}: [${symbolName}]\n`;
-    } else {
-      node = expr.properties[expr.properties.length - 1];
-      position = node.getEnd();
-      // Get the indentation of the last element, if any.
-      const text = node.getFullText(source);
-      if (text.match('^\r?\r?\n')) {
-        toInsert = `,${text.match(/^\r?\n\s+/)[0]}${metadataField}: [${symbolName}]`;
-      } else {
-        toInsert = `, ${metadataField}: [${symbolName}]`;
-      }
-    }
-  } else if (node.kind == ts.SyntaxKind.ArrayLiteralExpression) {
-    // We found the field but it's empty. Insert it just before the `]`.
-    position--;
-    toInsert = `${symbolName}`;
-  } else {
-    // Get the indentation of the last element, if any.
-    const text = node.getFullText(source);
-    if (text.match(/^\r?\n/)) {
-      toInsert = `,${text.match(/^\r?\n(\r?)\s+/)[0]}${symbolName}`;
-    } else {
-      toInsert = `, ${symbolName}`;
-    }
-  }
-  if (importPath !== null) {
-    return [
-      new InsertChange(ngModulePath, position, toInsert),
-      insertImport(source, ngModulePath, symbolName.replace(/\..*$/, ''), importPath)
-    ];
-  }
-
-  return [new InsertChange(ngModulePath, position, toInsert)];
+  return originalAddSymbolToNgModuleMetadata(source, ngModulePath, metadataField, symbolName, importPath);
 }
 
 /**
@@ -489,7 +133,7 @@ export function addDeclarationToModule(
   classifiedName: string,
   importPath: string
 ): Change[] {
-  return addSymbolToNgModuleMetadata(source, modulePath, 'declarations', classifiedName, importPath);
+  return originalAddDeclarationToModule(source, modulePath, classifiedName, importPath);
 }
 
 /**
@@ -501,7 +145,7 @@ export function addImportToModule(
   classifiedName: string,
   importPath: string
 ): Change[] {
-  return addSymbolToNgModuleMetadata(source, modulePath, 'imports', classifiedName, importPath);
+  return originalAddImportToModule(source, modulePath, classifiedName, importPath);
 }
 
 /**
@@ -513,7 +157,7 @@ export function addProviderToModule(
   classifiedName: string,
   importPath: string
 ): Change[] {
-  return addSymbolToNgModuleMetadata(source, modulePath, 'providers', classifiedName, importPath);
+  return originalAddProviderToModule(source, modulePath, classifiedName, importPath);
 }
 
 /**
@@ -525,7 +169,7 @@ export function addExportToModule(
   classifiedName: string,
   importPath: string
 ): Change[] {
-  return addSymbolToNgModuleMetadata(source, modulePath, 'exports', classifiedName, importPath);
+  return originalAddExportToModule(source, modulePath, classifiedName, importPath);
 }
 
 /**
@@ -537,7 +181,7 @@ export function addBootstrapToModule(
   classifiedName: string,
   importPath: string
 ): Change[] {
-  return addSymbolToNgModuleMetadata(source, modulePath, 'bootstrap', classifiedName, importPath);
+  return originalAddBootstrapToModule(source, modulePath, classifiedName, importPath);
 }
 
 /**
@@ -549,30 +193,26 @@ export function addEntryComponentToModule(
   classifiedName: string,
   importPath: string
 ): Change[] {
-  return addSymbolToNgModuleMetadata(source, modulePath, 'entryComponents', classifiedName, importPath);
+  return originalAddEntryComponentToModule(source, modulePath, classifiedName, importPath);
 }
 
 /**
  * Determine if an import already exists.
  */
 export function isImported(source: ts.SourceFile, classifiedName: string, importPath: string): boolean {
-  const allNodes = getSourceNodes(source);
-  const matchingNodes = allNodes
-    .filter(node => node.kind === ts.SyntaxKind.ImportDeclaration)
-    .filter((imp: ts.ImportDeclaration) => imp.moduleSpecifier.kind === ts.SyntaxKind.StringLiteral)
-    .filter((imp: ts.ImportDeclaration) => {
-      return (<ts.StringLiteral>imp.moduleSpecifier).text === importPath;
-    })
-    .filter((imp: ts.ImportDeclaration) => {
-      if (!imp.importClause) {
-        return false;
-      }
-      const nodes = findNodes(imp.importClause, ts.SyntaxKind.ImportSpecifier).filter(
-        n => n.getText() === classifiedName
-      );
+  return originalIsImported(source, classifiedName, importPath);
+}
 
-      return nodes.length > 0;
-    });
+/**
+ * Returns the RouterModule declaration from NgModule metadata, if any.
+ */
+export function getRouterModuleDeclaration(source: ts.SourceFile): ts.Expression | undefined {
+  return originalGetRouterModuleDeclaration(source);
+}
 
-  return matchingNodes.length > 0;
+/**
+ * Adds a new route declaration to a router module (i.e. has a RouterModule declaration)
+ */
+export function addRouteDeclarationToModule(source: ts.SourceFile, fileToAdd: string, routeLiteral: string): Change {
+  return originalAddRouteDeclarationToModule(source, fileToAdd, routeLiteral);
 }