+++ src/angular/latest-versions.ts
@@ -1,18 +1,3 @@
-/**
- * @license
- * Copyright Google Inc. All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
+import { latestVersions as originalLatestVersions } from '@schematics/angular/utility/latest-versions';
 
-export const latestVersions = {
-  // These versions should be kept up to date with latest Angular peer dependencies.
-  Angular: '^6.1.0',
-  RxJs: '^6.0.0',
-  ZoneJs: '~0.8.26',
-  TypeScript: '~2.9.2',
-  // The versions below must be manually updated when making a new devkit release.
-  DevkitBuildAngular: '~0.8.0-beta.0',
-  DevkitBuildNgPackagr: '~0.8.0-beta.0'
-};
+export const latestVersions = originalLatestVersions;