+++ src/material/get-project.ts
@@ -1,23 +1,10 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
-import { WorkspaceProject, WorkspaceSchema } from '../angular';
+import { getProjectFromWorkspace as originalGetProjectFromWorkspace } from '@angular/cdk/schematics';
+import { WorkspaceSchema, WorkspaceProject } from '@angular-devkit/core/src/experimental/workspace';
 
 /**
  * Finds the specified project configuration in the workspace. Throws an error if the project
  * couldn't be found.
  */
 export function getProjectFromWorkspace(workspace: WorkspaceSchema, projectName?: string): WorkspaceProject {
-  let project = workspace.projects[projectName || workspace.defaultProject];
-
-  if (!project) {
-    throw new Error(`Could not find project in workspace: ${projectName}`);
-  }
-
-  return project;
+  return originalGetProjectFromWorkspace(workspace, projectName);
 }